[FFM-5478]: Ruby SDK failed to cache some flags due to encoding
Remove the file cache as it is not required and caused bugs.
Ensure the segments are stored using set_segment instead of set_flag on startup
Add the latest evaluation tests

The Ruby SDK was failing to store some flags in its cache on startup,
this meant the SDK would not function as expected.
This only seemed to be a problem for the file cache, rather than in mem
cache.   The file cache could only handle UTF-8 data.

There was also a bug during startup where set_flag was being called in
place of set_segment, which meant no group data was stored in the cache.

The evaluation tests were included to help test and verify the above fixes.