FFM-7325 - Improve authentication retry logic
What
Harden the authentication retry logic to only retry on certain HTTP codes, certain error codes will be
treated as transient and others not. We also want to ensure that while we are authenticating that
default values will be served.

Why
We're doing a genernal review of network resiliency on auth and hardening code.

Testing
Manual + new unit tests added